Adapt the script to work with receiver-type=beast-tcp
If a user is not using dvb-t as receiver-type setting, the fr24 script
for installing dump1090 does not run.  In this case just install
dump1090-mutability directly.  Removing dump1090-mutability before
installing it should not be necessary and can be a problem if it was
previously configured for beast-tcp.  Also remove sudo for several
commands as the script should run as root anyway.